refactor(console-ui): split admin main.ts into per-feature modules
Break the 1022-line admin/operator console main.ts monolith into nine
sibling TypeScript modules so each feature has a single home and the
orchestrator stays small.

Module map:
- admin-types.ts        — shared type aliases + Window globals
- admin-state.ts        — encapsulated module-level mutable state
- admin-api.ts          — byId, authHeaders, api, slash command client
- admin-render.ts       — DOM writers, markdown/timing/source rendering
- admin-conversations.ts — conversation list, CRUD, context menu
- admin-query.ts        — query tab + SSE stream + slash dispatch
- admin-ingest.ts       — ingest tab + slash dispatch
- admin-health.ts       — health/log refresh
- admin-admin.ts        — admin tab API key handlers

main.ts is now a 56-line orchestrator that initializes tabs and binds
each feature on startup. Behavior is preserved exactly; closure-shared
state moved behind getter/setter functions in admin-state.

Imports use ".js" extensions so the compiled ES modules load directly
in the browser via <script type="module">.
