+++ server/console/static/admin-types.js
@@ -1,0 +1,10 @@
+/**
+ * @summary
+ * Shared TypeScript type aliases used across the admin/operator console modules.
+ * Exports: JsonObject, QueryResult, StreamEventData, TokenBudgetPayload, TimingPayload,
+ *   ConsoleCommandSpec, CommandExecution, ConversationMeta, ConversationTurn,
+ *   MarkedLike, DomPurifyLike
+ * Deps: none
+ * @end-summary
+ */
+export {};
